ec2: include attributes in instance observation, fix reconciling tags
This ensures that when tags change, the EC2 instance reconciles the
update. Previously the ec2 instance controller supported updating tags
via `CreateTags` but did not perform reconciliation when tags changed as
it was not included in the check for if the EC2 instance was up to date.

However, the way `Update` is currently implemented had the implicit
assumption that the EC2 instance must be stopped for all updates, which
is true for all the ModifyInstanceAttribute calls, however, tags can be
applied at runtime. This changes the logic of the controller to (1)
include the instance attributes in the instance observations, (2) only
attempt to modify instance attributes if the observation does not match
the spec.